pwsh: Cleanup powershell scripts (not complete)
This is the beginning of the xonsh scripts migration. The cleanup
is not complete yet, because there are some scripts that are
still not migrated, but in the actual status a project can be created
without the need of the powershell scripts.

The missing scripts will be migrated in the next commits, and are more
specific by template than in a general need for the workspace work
flow.

Signed-off-by: Matheus Castello <matheus.castello@toradex.com>