scripts: tasks: Add support for Docker in Docker from scratch
The Github actions already have support for Docker in Docker, but
in the way that the it uses the same Docker socket from the host
machine. For our tests the best was to have their own Docker socket
and all clean. For this purpose was added some edge cases where
the Docker host is set to the container IP and the port 2376, if it
detectes that this is needed.

Signed-off-by: Matheus Castello <matheus.castello@toradex.com>