github: workflow: Add CI for publish the torizon-utils package
The templates uses xonsh with Python 🤢 now 😢.
We need some utils to help us to create the templates and
the CI/CD pipeline, so I created a package called torizon-templates-utils.
This package will be published on PyPI every time that a workflow
finishes successfully and there is a new version.

Signed-off-by: Matheus Castello <matheus.castello@toradex.com>